netlink-socket: Check for nil sock in nl_sock_recv__()
Under certain circuntances of high load, the parameter 'sock' can be
passed as nil, causing a segmentation fault when it is dereferenced when
calling recvmsg().

The ovs-vswitchd.log will show something like this:

    fatal_signal(revalidator138)|WARN|terminating with signal 11 (signal 11)

Tested this change under the same circuntances that the segmentation
fault was found and it survived for four days in a row without any
issue.

Signed-off-by: Murilo Opsfelder Araujo <muriloo@linux.ibm.com>